timestamp: clarify that strptime produces DateTime not Duration
Prompted by: https://stackoverflow.com/questions/75654140/trouble-with-strptime-conversion-of-duration-time-string I went down a rabbit hole into how Dates and Times and Durations are handled.

The docs as of now are misleading, as DateTime cannot be safely cast to Time.

Maybe there should be a section on "Duration Stamp" parsing. The hacks suggested in the SO answer are not ideal. E.g. duration can certainly be negative. What about Time? Probably not. So Duration stamps should be handled separately. Otherwise all sorts of subtle bugs can creep in.

This PR is an attempt to defuse the ambiguity without proposing a clear way forward - this should be separately handled.